# bigbike-backend/src/main/resources/openapi/bigbike-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "BigBike Backend API",
    "version": "1.0.0",
    "description": "REST API for BigBike — Vietnam's premium motorcycle marketplace. Rewrite of WordPress backend. All endpoints return ApiDataResponse<T> or ApiListResponse<T> envelopes."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development"
    },
    {
      "url": "https://api.bigbike.vn",
      "description": "Production"
    }
  ],
...
  "paths": {
...
    "/api/v1/products/{productId}/reviews": {
      "get": {
        "tags": [
          "Public Catalog"
        ],
        "summary": "Get approved product reviews",
        "operationId": "getProductReviews",
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            }
          },
          {
            "name": "size",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10,
              "minimum": 1,
              "maximum": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Approved reviews for the requested product",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProductReviewsDataResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "post": {
        "tags": [
          "Public Catalog"
        ],
        "summary": "Submit product review",
        "operationId": "submitProductReview",
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitReviewRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Review submitted and pending moderation. Also returned (with no persistence) when honeypot is filled."
          },
          "400": {
            "description": "Validation error (missing/invalid authorName, rating, or comment)"
          },
          "404": {
            "description": "Product not found"
          },
          "409": {
            "description": "Duplicate review — same productId + normalized (trim/lowercase/collapsed-whitespace) authorName + comment submitted within the last 24 hours. Rating is not part of the duplicate key, and reviews in any moderation status (PENDING/APPROVED/SPAM/TRASH) are counted."
          },
          "429": {
            "description": "Rate limit exceeded — 5 submissions per IP per minute"
          }
        }
      }
    },
...
    "/api/v1/customer/orders": {
      "get": {
        "tags": [
          "Customer Orders"
        ],
        "summary": "List customer's own orders",
        "operationId": "listCustomerOrders",
        "security": [
          {
            "CustomerSession": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "size",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order list"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/v1/customer/orders/{orderId}": {
      "get": {
        "tags": [
          "Customer Orders"
        ],
        "summary": "Get customer order detail",
        "operationId": "getCustomerOrder",
        "security": [
          {
            "CustomerSession": []
          }
        ],
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order detail"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/api/v1/admin/orders/{orderId}": {
      "get": {
        "tags": [
          "Admin Orders"
        ],
        "summary": "Get admin order detail",
        "operationId": "adminGetOrder",
        "security": [
          {
            "AdminBearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order detail"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/api/v1/admin/customers/{customerId}": {
      "get": {
        "tags": [
          "Admin Customers"
        ],
        "summary": "Get customer by id (admin)",
        "operationId": "adminGetCustomer",
        "security": [
          {
            "AdminBearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "customerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Customer detail"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "patch": {
        "tags": [
          "Admin Customers"
        ],
        "summary": "Update customer (admin)",
        "operationId": "adminUpdateCustomer",
        "security": [
          {
            "AdminBearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AdminBearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Admin JWT access token. Obtain via POST /api/v1/auth/login. Header: Authorization: Bearer <accessToken>"
      },
      "CustomerSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "bb_session",
        "description": "Customer session cookie. Set automatically after POST /api/v1/customer/auth/login."
      },
      "CsrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "CSRF token required for all cart/checkout mutations. Value must match the bb_csrf cookie set by the server."
      }
    },
    "responses": {
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiErrorResponse"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Authentication required or token expired",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiErrorResponse"
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "ApiMeta": {
        "type": "object",
        "properties": {
          "requestId": {
            "type": "string",
            "format": "uuid"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer"
          },
          "pageSize": {
            "type": "integer"
          },
          "totalItems": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer"
          },
          "hasNext": {
            "type": "boolean"
          },
          "hasPrevious": {
            "type": "boolean"
          }
        }
      },
...
      "PublicReviewItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "authorName": {
            "type": "string"
          },
          "rating": {
            "type": "integer"
          },
          "comment": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "authorAvatarUrl": {
            "type": "string",
            "nullable": true,
            "description": "Linked customer's current avatar URL when the review was submitted while logged in; resolved live at read time (not a snapshot). Null for guest reviews or when the customer has no avatar."
          }
        }
      },
      "PublicProductReviewsResponse": {
        "type": "object",
        "properties": {
          "avgRating": {
            "type": "number",
            "format": "double"
          },
          "totalReviews": {
            "type": "integer",
            "format": "int64"
          },
          "reviews": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublicReviewItem"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMeta"
          }
        }
      },
      "PublicProductReviewsDataResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/PublicProductReviewsResponse"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        }
      },
      "SubmitReviewRequest": {
        "type": "object",
        "required": [
          "authorName",
          "rating"
        ],
        "properties": {
          "authorName": {
            "type": "string",
            "maxLength": 80
          },
          "rating": {
            "type": "integer",
            "minimum": 1,
            "maximum": 5
          },
          "comment": {
            "type": "string",
            "maxLength": 1000
          },
          "website": {
            "type": "string",
            "description": "Honeypot field — real clients leave this empty/omitted. Non-empty value (after trim) is silently dropped server-side: response is success-shaped 201 but no review is persisted.",
            "example": ""
          }
        }
      },
...
      "ApiErrorResponse": {
        "type": "object",
        "description": "Standard error envelope used by all error responses",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiError"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        }
      },
      "ApiError": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "VALIDATION_ERROR"
          },
          "message": {
            "type": "string",
            "example": "Validation failed."
          },
          "details": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApiErrorDetail"
            }
          }
        }
      },
      "ApiErrorDetail": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      }
    }
  }
}